KYLIN-2783 refactor CuboidScheduler
