+++ core-common/src/main/resources/kylin-defaults.properties
@@ -145,6 +145,8 @@ kylin.engine.mr.uhc-reducer-count=1
 
 ### CUBE | DICTIONARY ###
 
+kylin.cube.cuboid-scheduler=org.apache.kylin.cube.cuboid.DefaultCuboidScheduler
+
 # 'auto', 'inmem', 'layer' or 'random' for testing 
 kylin.cube.algorithm=layer
